Revert "Merge branch '1.3' into 1.4"
This reverts commit 72e764841e638c4e619c1dd9cd6b271eea9eba0c, reversing
changes made to e1479347890684b702f2d2c6937f343a96b51cdc.

+++ nifty-controls/src/main/java/de/lessvoid/nifty/controls/scrollbar/ScrollbarImpl.java
@@ -197,7 +197,7 @@ public class ScrollbarImpl {
     if (worldMax == 0.f) {
       return 0.f;
     }
-    return Math.round(worldValue / worldMaxValue * viewMaxValue);
+    return (int) Math.round(worldValue / worldMaxValue * viewMaxValue);
   }
 
   private void ensureWorldPageSize() {
